Update configuration files
