--- app/models/spree/product_decorator.rb
@@ -1,1 +1,0 @@
-Spree::Product.include(SolidusFavorites::ProductDecorator)
