--- app/models/spree/user_decorator.rb
@@ -1,1 +1,0 @@
-Spree.user_class.include(SolidusFavorites::UserDecorator)
