+++ spec/spec_helper.rb
@@ -1,17 +1,24 @@
 # frozen_string_literal: true
 
+# Configure Rails Environment
 ENV['RAILS_ENV'] = 'test'
 
-require File.expand_path('../dummy/config/environment.rb', __FILE__)
+# Run Coverage report
+require 'solidus_dev_support/rspec/coverage'
 
-require 'rspec/rails'
+require File.expand_path('dummy/config/environment.rb', __dir__)
 
-Dir[File.join(File.dirname(__FILE__), '/support/**/*.rb')].each { |file| require file }
+# Requires factories and other useful helpers defined in spree_core.
+require 'solidus_dev_support/rspec/feature_helper'
 
+# Requires supporting ruby files with custom matchers and macros, etc,
+# in spec/support/ and its subdirectories.
+Dir[File.join(File.dirname(__FILE__), 'support/**/*.rb')].each { |f| require f }
+
+# Requires factories defined in lib/solidus_favorites/factories.rb
 require 'solidus_favorites/factories'
-require 'spree/testing_support/controller_requests'
 
 RSpec.configure do |config|
-  config.include Spree::TestingSupport::ControllerRequests, type: :controller
-  config.extend Spree::TestingSupport::AuthorizationHelpers, type: :feature
+  config.infer_spec_type_from_file_location!
+  config.use_transactional_fixtures = false
 end
