--- spec/support/database_cleaner.rb
@@ -1,16 +1,0 @@
-require 'database_cleaner'
-
-RSpec.configure do |config|
-  config.before(:suite) do
-    DatabaseCleaner.clean_with :truncation
-  end
-
-  config.before do
-    DatabaseCleaner.strategy = :transaction
-    DatabaseCleaner.start
-  end
-
-  config.after do
-    DatabaseCleaner.clean
-  end
-end
