--- spec/support/factory_bot.rb
@@ -1,5 +1,0 @@
-require 'factory_bot'
-
-RSpec.configure do |config|
-  config.include FactoryBot::Syntax::Methods
-end
